backend/requirements.txt
langgraph>=0.2.0
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
fastapi>=0.109.0
uvicorn>=0.27.0
python-dotenv>=1.0.0
weaviate-client>=4.5.0
pydantic>=2.0.0
websockets>=12.0
copilotkit>=0.1.39
